Add checkbox post-processor
This post processor is a python script
that should be ran on each html file after "mdbook build" runs.

The script surgically replaces basic list items from the ingredients
and the methods, and converts them in an HTML checkbox, which changes
the text properties when clicked.

The javascript handles the actions that should be taken when clicked.